moving WorkflowRegtest to a separate file
